feat!: Move InnerProduct to new directory
