+++ crates/proof-of-sql/src/base/commitment/mod.rs
@@ -6,7 +6,6 @@ pub use blitzar::{
     proof::InnerProductProof,
 };
 use core::ops::{AddAssign, SubAssign};
-use curve25519_dalek::ristretto::RistrettoPoint;
 
 mod committable_column;
 pub(crate) use committable_column::CommittableColumn;
@@ -15,7 +14,6 @@ mod vec_commitment_ext;
 pub use vec_commitment_ext::{NumColumnsMismatch, VecCommitmentExt};
 
 mod column_bounds;
-use super::scalar::Curve25519Scalar;
 pub use column_bounds::{Bounds, ColumnBounds, NegativeBounds};
 
 mod column_commitment_metadata;
@@ -78,43 +76,6 @@ pub trait Commitment:
         offset: usize,
         setup: &Self::PublicSetup<'_>,
     );
-}
-
-impl Commitment for RistrettoPoint {
-    type Scalar = Curve25519Scalar;
-    type PublicSetup<'a> = ();
-    #[cfg(feature = "blitzar")]
-    fn compute_commitments(
-        commitments: &mut [Self],
-        committable_columns: &[CommittableColumn],
-        offset: usize,
-        _setup: &Self::PublicSetup<'_>,
-    ) {
-        let sequences = Vec::from_iter(committable_columns.iter().map(Into::into));
-        let mut compressed_commitments = vec![Default::default(); committable_columns.len()];
-        blitzar::compute::compute_curve25519_commitments(
-            &mut compressed_commitments,
-            &sequences,
-            offset as u64,
-        );
-        commitments
-            .iter_mut()
-            .zip(compressed_commitments.iter())
-            .for_each(|(c, cc)| {
-                *c = cc.decompress().expect(
-                    "invalid ristretto point decompression in Commitment::compute_commitments",
-                );
-            });
-    }
-    #[cfg(not(feature = "blitzar"))]
-    fn compute_commitments(
-        _commitments: &mut [Self],
-        _committable_columns: &[CommittableColumn],
-        _offset: usize,
-        _setup: &Self::PublicSetup<'_>,
-    ) {
-        unimplemented!()
-    }
 }
 
 mod commitment_evaluation_proof;
