+++ crates/proof-of-sql/src/base/scalar/mod.rs
@@ -5,7 +5,6 @@ mod mont_scalar;
 #[cfg(test)]
 mod mont_scalar_test;
 use core::{cmp::Ordering, ops::Sub};
-pub use mont_scalar::Curve25519Scalar;
 pub(crate) use mont_scalar::MontScalar;
 mod mont_scalar_from;
 #[cfg(test)]
