+++ crates/proof-of-sql/src/base/scalar/mont_scalar.rs
@@ -115,11 +115,6 @@ impl<T: MontConfig<4>> Ord for MontScalar<T> {
 // end replacement for #[derive(...)]
 // --------------------------------------------------------------------------------
 
-/// A wrapper type around the field element `ark_curve25519::Fr` and should be used in place of `ark_curve25519::Fr`.
-///
-/// Using the `Scalar` trait rather than this type is encouraged to allow for easier switching of the underlying field.
-pub type Curve25519Scalar = MontScalar<ark_curve25519::FrConfig>;
-
 impl<T: MontConfig<4>> MontScalar<T> {
     /// Convenience function for creating a new `MontScalar<T>` from the underlying `Fp256<MontBackend<T, 4>>`. Should only be used in tests.
     #[cfg(test)]
@@ -139,12 +134,12 @@ impl<T: MontConfig<4>> MontScalar<T> {
     pub fn to_bytes_le(&self) -> Vec<u8> {
         self.0.into_bigint().to_bytes_le()
     }
-    /// Convenience function for converting a slice of `ark_curve25519::Fr` into a vector of `Curve25519Scalar`. Should not be used outside of tests.
+    /// Convenience function for converting a slice of `ark_curve25519::Fr` into a vector of `TestScalar`. Should not be used outside of tests.
     #[cfg(test)]
     pub fn wrap_slice(slice: &[Fp256<MontBackend<T, 4>>]) -> Vec<Self> {
         slice.iter().copied().map(Self).collect()
     }
-    /// Convenience function for converting a slice of `Curve25519Scalar` into a vector of `ark_curve25519::Fr`. Should not be used outside of tests.
+    /// Convenience function for converting a slice of `TestScalar` into a vector of `ark_curve25519::Fr`. Should not be used outside of tests.
     #[cfg(test)]
     pub fn unwrap_slice(slice: &[Self]) -> Vec<Fp256<MontBackend<T, 4>>> {
         slice.iter().map(|x| x.0).collect()
@@ -194,31 +189,6 @@ impl<T: MontConfig<4>> ark_std::UniformRand for MontScalar<T> {
     }
 }
 
-impl core::ops::Mul<curve25519_dalek::ristretto::RistrettoPoint> for Curve25519Scalar {
-    type Output = curve25519_dalek::ristretto::RistrettoPoint;
-    fn mul(self, rhs: curve25519_dalek::ristretto::RistrettoPoint) -> Self::Output {
-        curve25519_dalek::scalar::Scalar::from(self) * rhs
-    }
-}
-impl core::ops::Mul<Curve25519Scalar> for curve25519_dalek::ristretto::RistrettoPoint {
-    type Output = curve25519_dalek::ristretto::RistrettoPoint;
-    fn mul(self, rhs: Curve25519Scalar) -> Self::Output {
-        self * curve25519_dalek::scalar::Scalar::from(rhs)
-    }
-}
-impl core::ops::Mul<&curve25519_dalek::ristretto::RistrettoPoint> for Curve25519Scalar {
-    type Output = curve25519_dalek::ristretto::RistrettoPoint;
-    fn mul(self, rhs: &curve25519_dalek::ristretto::RistrettoPoint) -> Self::Output {
-        curve25519_dalek::scalar::Scalar::from(self) * rhs
-    }
-}
-impl core::ops::Mul<Curve25519Scalar> for &curve25519_dalek::ristretto::RistrettoPoint {
-    type Output = curve25519_dalek::ristretto::RistrettoPoint;
-    fn mul(self, rhs: Curve25519Scalar) -> Self::Output {
-        self * curve25519_dalek::scalar::Scalar::from(rhs)
-    }
-}
-
 impl<'a, T: MontConfig<4>> Sum<&'a Self> for MontScalar<T> {
     fn sum<I: Iterator<Item = &'a Self>>(iter: I) -> Self {
         Self(iter.map(|x| x.0).sum())
@@ -264,18 +234,6 @@ impl<T: MontConfig<4>> core::ops::Neg for &MontScalar<T> {
     type Output = MontScalar<T>;
     fn neg(self) -> Self::Output {
         MontScalar(-self.0)
-    }
-}
-impl From<Curve25519Scalar> for curve25519_dalek::scalar::Scalar {
-    fn from(value: Curve25519Scalar) -> Self {
-        (&value).into()
-    }
-}
-
-impl From<&Curve25519Scalar> for curve25519_dalek::scalar::Scalar {
-    fn from(value: &Curve25519Scalar) -> Self {
-        let bytes = ark_ff::BigInteger::to_bytes_le(&value.0.into_bigint());
-        curve25519_dalek::scalar::Scalar::from_canonical_bytes(bytes.try_into().unwrap()).unwrap()
     }
 }
 
@@ -354,15 +312,6 @@ impl<T: MontConfig<4>> Display for MontScalar<T> {
             }
         }
     }
-}
-
-impl super::Scalar for Curve25519Scalar {
-    const MAX_SIGNED: Self = Self(ark_ff::MontFp!(
-        "3618502788666131106986593281521497120428558179689953803000975469142727125494"
-    ));
-    const ZERO: Self = Self(ark_ff::MontFp!("0"));
-    const ONE: Self = Self(ark_ff::MontFp!("1"));
-    const TWO: Self = Self(ark_ff::MontFp!("2"));
 }
 
 impl<T> TryFrom<MontScalar<T>> for bool
