+++ crates/proof-of-sql/src/base/slice_ops/inner_product_test.rs
@@ -1,5 +1,6 @@
+use crate::base::scalar::test_scalar::TestScalar;
+
 use super::*;
-use crate::base::scalar::Curve25519Scalar;
 
 #[test]
 fn test_inner_product() {
@@ -19,39 +20,39 @@ fn test_inner_product_different_lengths() {
 /// test inner producr with scalar
 #[test]
 fn test_inner_product_scalar() {
-    let a = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
-    let b = vec![Curve25519Scalar::from(2u64), Curve25519Scalar::from(3u64)];
-    assert_eq!(Curve25519Scalar::from(8u64), inner_product(&a, &b));
+    let a = vec![TestScalar::from(1u64), TestScalar::from(2u64)];
+    let b = vec![TestScalar::from(2u64), TestScalar::from(3u64)];
+    assert_eq!(TestScalar::from(8u64), inner_product(&a, &b));
 }
 
 /// test uneven inner product with scalars
 #[test]
 fn test_inner_product_scalar_uneven() {
-    let a = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
+    let a = vec![TestScalar::from(1u64), TestScalar::from(2u64)];
     let b = vec![
-        Curve25519Scalar::from(2u64),
-        Curve25519Scalar::from(3u64),
-        Curve25519Scalar::from(4u64),
+        TestScalar::from(2u64),
+        TestScalar::from(3u64),
+        TestScalar::from(4u64),
     ];
-    assert_eq!(Curve25519Scalar::from(8u64), inner_product(&a, &b));
+    assert_eq!(TestScalar::from(8u64), inner_product(&a, &b));
 }
 
-/// test inner product with curve25519scalar
+/// test inner product with TestScalar
 #[test]
-fn test_inner_product_curve25519scalar() {
-    let a = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
-    let b = vec![Curve25519Scalar::from(2u64), Curve25519Scalar::from(3u64)];
-    assert_eq!(Curve25519Scalar::from(8u64), inner_product(&a, &b));
+fn test_inner_product_testscalar() {
+    let a = vec![TestScalar::from(1u64), TestScalar::from(2u64)];
+    let b = vec![TestScalar::from(2u64), TestScalar::from(3u64)];
+    assert_eq!(TestScalar::from(8u64), inner_product(&a, &b));
 }
 
-/// test uneven inner product with curve25519scalars
+/// test uneven inner product with TestScalars
 #[test]
-fn test_inner_product_curve25519scalar_uneven() {
-    let a = vec![Curve25519Scalar::from(1u64), Curve25519Scalar::from(2u64)];
+fn test_inner_product_testscalar_uneven() {
+    let a = vec![TestScalar::from(1u64), TestScalar::from(2u64)];
     let b = vec![
-        Curve25519Scalar::from(2u64),
-        Curve25519Scalar::from(3u64),
-        Curve25519Scalar::from(4u64),
+        TestScalar::from(2u64),
+        TestScalar::from(3u64),
+        TestScalar::from(4u64),
     ];
-    assert_eq!(Curve25519Scalar::from(8u64), inner_product(&a, &b));
+    assert_eq!(TestScalar::from(8u64), inner_product(&a, &b));
 }
