+++ crates/proof-of-sql/src/proof_primitive/inner_product/curve_25519_scalar.rs
@@ -1,0 +1,53 @@
+use crate::base::scalar::{MontScalar, Scalar};
+use ark_ff::PrimeField;
+
+/// A wrapper type around the field element `ark_curve25519::Fr` and should be used in place of `ark_curve25519::Fr`.
+///
+/// Using the `Scalar` trait rather than this type is encouraged to allow for easier switching of the underlying field.
+pub type Curve25519Scalar = MontScalar<ark_curve25519::FrConfig>;
+
+impl core::ops::Mul<curve25519_dalek::ristretto::RistrettoPoint> for Curve25519Scalar {
+    type Output = curve25519_dalek::ristretto::RistrettoPoint;
+    fn mul(self, rhs: curve25519_dalek::ristretto::RistrettoPoint) -> Self::Output {
+        curve25519_dalek::scalar::Scalar::from(self) * rhs
+    }
+}
+impl core::ops::Mul<Curve25519Scalar> for curve25519_dalek::ristretto::RistrettoPoint {
+    type Output = curve25519_dalek::ristretto::RistrettoPoint;
+    fn mul(self, rhs: Curve25519Scalar) -> Self::Output {
+        self * curve25519_dalek::scalar::Scalar::from(rhs)
+    }
+}
+impl core::ops::Mul<&curve25519_dalek::ristretto::RistrettoPoint> for Curve25519Scalar {
+    type Output = curve25519_dalek::ristretto::RistrettoPoint;
+    fn mul(self, rhs: &curve25519_dalek::ristretto::RistrettoPoint) -> Self::Output {
+        curve25519_dalek::scalar::Scalar::from(self) * rhs
+    }
+}
+impl core::ops::Mul<Curve25519Scalar> for &curve25519_dalek::ristretto::RistrettoPoint {
+    type Output = curve25519_dalek::ristretto::RistrettoPoint;
+    fn mul(self, rhs: Curve25519Scalar) -> Self::Output {
+        self * curve25519_dalek::scalar::Scalar::from(rhs)
+    }
+}
+impl From<Curve25519Scalar> for curve25519_dalek::scalar::Scalar {
+    fn from(value: Curve25519Scalar) -> Self {
+        (&value).into()
+    }
+}
+
+impl From<&Curve25519Scalar> for curve25519_dalek::scalar::Scalar {
+    fn from(value: &Curve25519Scalar) -> Self {
+        let bytes = ark_ff::BigInteger::to_bytes_le(&value.0.into_bigint());
+        curve25519_dalek::scalar::Scalar::from_canonical_bytes(bytes.try_into().unwrap()).unwrap()
+    }
+}
+
+impl Scalar for Curve25519Scalar {
+    const MAX_SIGNED: Self = Self(ark_ff::MontFp!(
+        "3618502788666131106986593281521497120428558179689953803000975469142727125494"
+    ));
+    const ZERO: Self = Self(ark_ff::MontFp!("0"));
+    const ONE: Self = Self(ark_ff::MontFp!("1"));
+    const TWO: Self = Self(ark_ff::MontFp!("2"));
+}
