+++ crates/proof-of-sql/src/proof_primitive/mod.rs
@@ -1,3 +1,5 @@
 //! TODO: add docs
 pub mod dory;
+/// Module for the InnerProductProof code
+pub mod inner_product;
 pub(crate) mod sumcheck;
