+++ crates/proof-of-sql/src/sql/proof/verification_builder_test.rs
@@ -1,5 +1,6 @@
+use crate::proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar;
+
 use super::{SumcheckMleEvaluations, VerificationBuilder};
-use crate::base::scalar::Curve25519Scalar;
 use curve25519_dalek::ristretto::RistrettoPoint;
 use num_traits::Zero;
 use rand_core::OsRng;
