+++ crates/proof-of-sql/src/sql/proof_exprs/equals_expr_test.rs
@@ -2,8 +2,9 @@ use crate::{
     base::{
         commitment::InnerProductProof,
         database::{owned_table_utility::*, Column, OwnedTable, OwnedTableTestAccessor},
-        scalar::{Curve25519Scalar, Scalar},
+        scalar::Scalar,
     },
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
         proof_exprs::{test_utility::*, DynProofExpr, ProofExpr},
