+++ crates/proof-of-sql/src/sql/proof_plans/group_by_exec_test.rs
@@ -3,8 +3,8 @@ use crate::{
     base::{
         commitment::InnerProductProof,
         database::{owned_table_utility::*, OwnedTableTestAccessor, TestAccessor},
-        scalar::Curve25519Scalar,
     },
+    proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar,
     sql::{
         proof::{exercise_verification, VerifiableQueryResult},
         proof_exprs::test_utility::*,
