+++ crates/proof-of-sql/tests/decimal_integration_tests.rs
@@ -2,7 +2,8 @@
 #[cfg(feature = "blitzar")]
 use blitzar::proof::InnerProductProof;
 #[cfg(feature = "blitzar")]
-use proof_of_sql::base::{database::owned_table_utility::*, scalar::Curve25519Scalar as S};
+use proof_of_sql::base::database::owned_table_utility::*;
+use proof_of_sql::proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar as S;
 use proof_of_sql::sql::postprocessing::apply_postprocessing_steps;
 
 #[cfg(feature = "blitzar")]
@@ -60,7 +61,8 @@ fn run_query(
 #[cfg(feature = "blitzar")]
 mod decimal_query_tests {
     use crate::run_query;
-    use proof_of_sql::base::scalar::{Curve25519Scalar as S, Scalar};
+    use proof_of_sql::base::scalar::Scalar;
+    use proof_of_sql::proof_primitive::inner_product::curve_25519_scalar::Curve25519Scalar as S;
 
     #[test]
     fn we_can_query_decimals_exactly_matching_db_data() {
